Decouple two tests that could affect each other
